Moved tests from ModelsActionTest to ModelTest.
